requirements.txt
# Core Data & APIs
yfinance==0.2.28
alpha-vantage==2.3.1
requests==2.31.0
pandas==2.1.4
numpy==1.24.3

# Web Framework & Real-time
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
jinja2==3.1.2
python-multipart==0.0.6

# Data Analysis & AI
scikit-learn==1.3.2
ta-lib==0.4.28
scipy==1.11.4

# Visualization
plotly==5.17.0
matplotlib==3.8.2
seaborn==0.13.0

# Astrology & Spiritual APIs
pyephem==9.99
swisseph==2.10.03.2
timezonefinder==6.2.0

# UI & Styling
python-dotenv==1.0.0
colorama==0.4.6

# Utilities
schedule==1.2.0
python-dateutil==2.8.2
pytz==2023.3 